Refactor session storage implementation and update import paths
