+++ src/components/releases/edit/index.tsx
@@ -5,7 +5,7 @@ import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs'
 import { useNavigate, useParams } from 'react-router-dom'
 import { Release } from '../../../types/releases/release'
 import './styles.scss'
-import { SessionStorage } from '../../../utilities/git-api/local-storage/session-storage'
+import { SessionStorage } from '../../../utilities/git-api/storage/session-storage'
 import { useEffect, useState } from 'react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faCodeBranch, faCodeMerge, faCodePullRequest, faX } from '@fortawesome/free-solid-svg-icons'
