+++ src/components/releases/index.tsx
@@ -1,7 +1,7 @@
 import { AddTask, Edit, Launch } from '@mui/icons-material'
 import { Box, Button, Card, CardHeader, IconButton, Typography } from '@mui/material'
 import { Link, NavLink } from 'react-router-dom'
-import { SessionStorage } from '../../utilities/git-api/local-storage/session-storage'
+import { SessionStorage } from '../../utilities/git-api/storage/session-storage'
 import { Release } from '../../types/releases/release'
 
 export const Releases = () => {
